requirements.txt
pytest==7.4.0
selenium==4.15.0
requests==2.31.0
allure-pytest==2.13.2
webdriver-manager==4.0.0
python-dotenv==1.0.0
flake8==6.1.0
black==23.11.0